backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3
httpx==0.27.2
python-jose==3.3.0
google-generativeai==0.8.3
orjson==3.10.7
python-multipart==0.0.9

# FAISS and Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.3.1
numpy==1.24.3
redis>=5.0.0
redisvl>=0.0.1
